requirements.txt
-i https://pypi.org/simple
django==2.1.1
pycatalog==1.2.0
jinja2==2.10
django-environ==0.4.5
psycopg2-binary==2.7.5
django-extensions==2.1.3
django-tagulous==0.13.2
django-jinja==2.4.1
django-crispy-forms==1.7.2
